requirements.txt
# requirements.txt

# Manipulation de données
pandas
numpy

# Visualisation
matplotlib
seaborn
plotly

# Modélisation
scikit-learn
xgboost
tensorflow
keras

# Séries temporelles
statsmodels
prophet
tsfresh

# Web/API
requests
pyesios
python-dotenv
pyyaml
pyaemet
holidays
meteostat

# Dashboard
streamlit
dash

# Divers
tqdm
